added specifications on products route, and cleared about leads to alessandro update
