+++ node/clients/leadByEmail.ts
@@ -14,7 +14,6 @@ export default class LeadByEmail extends ExternalClient {
   constructor(context: IOContext, options?: InstanceOptions) {
     super('http://g0deojz10k.execute-api.us-east-2.amazonaws.com', context, {
       ...options,
-      // https://github.com/vtex-apps/yotpo-integration/blob/bf52a9af95fdf0160d5c804218fcc0691bdeec02/node/clients/yotpo.ts
       headers: {
         'X-VTEX-Use-Https': 'true',
         'Proxy-Authorization': context.authToken,
@@ -28,14 +27,4 @@ export default class LeadByEmail extends ExternalClient {
     })
   }
 
-  // public async getLeadWithHeaders(ctx: Context, email: string): Promise<IOResponse<string>> {
-  //   console.log("authToken:", ctx.vtex.authToken)
-  //   return this.http.getRaw(`/lead/${email}`, {
-  //     metric: 'leads-get-raw',
-  //     headers: {
-  //       "X-VTEX-Use-Https": true,
-  //       "Proxy-Authorization": ctx.vtex.authToken,
-  //     }
-  //   })
-  // }
 }
